+++ micronaut-tao/src/test/java/io/github/xezzon/tao/jpa/JpaWrapperTest.java
@@ -3,19 +3,14 @@ package io.github.xezzon.tao.jpa;
 import cn.hutool.core.util.IdUtil;
 import cn.hutool.core.util.RandomUtil;
 import cn.hutool.core.util.StrUtil;
+import io.github.xezzon.tao.jpa.user.GenderEnum;
+import io.github.xezzon.tao.jpa.user.User;
+import io.github.xezzon.tao.jpa.user.UserDAO;
+import io.github.xezzon.tao.jpa.user.UserRepository;
 import io.github.xezzon.tao.retrieval.CommonQuery;
-import io.micronaut.data.annotation.Repository;
-import io.micronaut.data.jpa.repository.JpaRepository;
 import io.micronaut.data.model.Page;
-import io.micronaut.runtime.event.ApplicationStartupEvent;
-import io.micronaut.runtime.event.annotation.EventListener;
 import io.micronaut.test.extensions.junit5.annotation.MicronautTest;
 import jakarta.inject.Inject;
-import jakarta.inject.Singleton;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
 import java.math.BigDecimal;
 import java.time.Instant;
 import java.time.LocalDate;
@@ -29,24 +24,68 @@ import java.util.List;
 import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.experimental.Accessors;
 import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.api.TestInstance.Lifecycle;
 
 /**
  * @author xezzon
  */
 @MicronautTest
+@TestInstance(Lifecycle.PER_CLASS)
 class JpaWrapperTest {
+
+  private static final List<User> DATASET = new ArrayList<>();
 
   @Inject
   protected UserDAO userDAO;
+  @Inject
+  protected UserRepository repository;
+
+  @BeforeAll
+  void beforeAll() {
+    for (int i = 0; i < Byte.MAX_VALUE; i++) {
+      User user = new User();
+      user.setId(IdUtil.getSnowflakeNextIdStr());
+      user.setName(RandomUtil.randomString(6));
+      user.setAge(RandomUtil.randomInt(6, 60));
+      user.setCredit(RandomUtil.randomBigDecimal(new BigDecimal("100.000")));
+      user.setGender(RandomUtil.randomEle(GenderEnum.values()));
+      user.setCreateTime(Instant.EPOCH.plus(
+          RandomUtil.randomLong(365 * 30, 365 * 300),
+          ChronoUnit.DAYS
+      ));
+      user.setDeleteDateTime(RandomUtil.randomBoolean() ?
+          LocalDateTime.of(
+              RandomUtil.randomInt(2000, 2999),
+              RandomUtil.randomInt(1, 11),
+              RandomUtil.randomInt(1, 27),
+              RandomUtil.randomInt(1, 11),
+              RandomUtil.randomInt(1, 59),
+              RandomUtil.randomInt(1, 59)
+          ) : null
+      );
+      user.setDeleted(user.getDeleteDateTime() != null);
+      user.setDeleteDate(
+          user.getDeleteDateTime() == null
+              ? null
+              : user.getDeleteDateTime().toLocalDate()
+      );
+      user.setDeleteTime(
+          user.getDeleteDateTime() == null
+              ? null
+              : user.getDeleteDateTime().toLocalTime()
+      );
+      DATASET.add(user);
+    }
+    repository.saveAll(DATASET);
+  }
 
   @Test
   void update() {
-    User user = UserDataset.getDataset().parallelStream()
+    User user = DATASET.parallelStream()
         .filter(o -> Objects.nonNull(o.getDeleteDateTime()))
         .filter(o -> !Objects.equals(o.getGender(), GenderEnum.UNKNOWN))
         .findAny().get();
@@ -90,7 +129,7 @@ class JpaWrapperTest {
   @Test
   void query() {
     String condition = "(name LLIKE 'J' OR (age GT 18)) AND (gender IN 'MALE' OR deleteDateTime NULL true)";
-    List<User> excepts = UserDataset.getDataset().parallelStream()
+    List<User> excepts = DATASET.parallelStream()
         .filter(user -> user.getName().startsWith("J") || user.getAge() > 18)
         .filter(user -> Objects.equals(GenderEnum.MALE, user.getGender())
             || user.getDeleteDateTime() == null)
@@ -118,8 +157,8 @@ class JpaWrapperTest {
   @Test
   void query_unpaged() {
     Page<User> users = userDAO.query(new CommonQuery());
-    Assertions.assertEquals(UserDataset.getDataset().size(), users.getTotalSize());
-    Assertions.assertEquals(UserDataset.getDataset().size(), users.getContent().size());
+    Assertions.assertEquals(DATASET.size(), users.getTotalSize());
+    Assertions.assertEquals(DATASET.size(), users.getContent().size());
   }
 
   /**
@@ -138,14 +177,14 @@ class JpaWrapperTest {
   @Test
   void query_string() {
     CommonQuery eqQuery = new CommonQuery();
-    String exceptName = UserDataset.getDataset().parallelStream()
+    String exceptName = DATASET.parallelStream()
         .map(User::getName)
         .findAny()
         .get();
     eqQuery.setFilter(String.format("name EQ '%s'", exceptName));
     Page<User> eqPage = userDAO.query(eqQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> Objects.equals(user.getName(), exceptName))
             .count(),
         eqPage.getTotalSize()
@@ -155,13 +194,13 @@ class JpaWrapperTest {
     neQuery.setFilter(String.format("name NE '%s'", exceptName));
     Page<User> nePage = userDAO.query(neQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> !Objects.equals(user.getName(), exceptName))
             .count(),
         nePage.getTotalSize()
     );
     Assertions.assertEquals(
-        UserDataset.getDataset().size(),
+        DATASET.size(),
         eqPage.getTotalSize() + nePage.getTotalSize()
     );
 
@@ -169,13 +208,13 @@ class JpaWrapperTest {
     llikeQuery.setFilter(String.format("name LLIKE '%s'", exceptName.substring(0, 3)));
     Page<User> llikePage = userDAO.query(llikeQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> user.getName().startsWith(exceptName.substring(0, 3)))
             .count(),
         llikePage.getTotalSize()
     );
 
-    Set<String> exceptId = UserDataset.getDataset().parallelStream()
+    Set<String> exceptId = DATASET.parallelStream()
         .limit(10)
         .map(User::getId)
         .collect(Collectors.toSet());
@@ -183,7 +222,7 @@ class JpaWrapperTest {
     inQuery.setFilter(String.format("id IN '%s'", StrUtil.join(",", exceptId)));
     Page<User> inPage = userDAO.query(inQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> exceptId.contains(user.getId()))
             .count(),
         inPage.getTotalSize()
@@ -193,13 +232,13 @@ class JpaWrapperTest {
     outQuery.setFilter(String.format("id OUT '%s'", StrUtil.join(",", exceptId)));
     Page<User> outPage = userDAO.query(outQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> !exceptId.contains(user.getId()))
             .count(),
         outPage.getTotalSize()
     );
     Assertions.assertEquals(
-        UserDataset.getDataset().size(),
+        DATASET.size(),
         inPage.getTotalSize() + outPage.getTotalSize()
     );
 
@@ -215,7 +254,7 @@ class JpaWrapperTest {
     emptyInQuery2.setFilter(String.format("name IN ',%s, ,'", exceptName));
     Page<User> emptyInPage2 = userDAO.query(emptyInQuery2);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> Objects.equals(user.getName(), exceptName))
             .count(),
         emptyInPage2.getTotalSize()
@@ -228,7 +267,7 @@ class JpaWrapperTest {
     inQuery.setFilter("gender IN 'MALE,FEMALE'");
     Page<User> inPage = userDAO.query(inQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> Set.of(GenderEnum.MALE, GenderEnum.FEMALE).contains(user.getGender()))
             .count(),
         inPage.getTotalSize()
@@ -238,13 +277,13 @@ class JpaWrapperTest {
     outQuery.setFilter("gender OUT 'MALE,FEMALE'");
     Page<User> outPage = userDAO.query(outQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> !Set.of(GenderEnum.MALE, GenderEnum.FEMALE).contains(user.getGender()))
             .count(),
         outPage.getTotalSize()
     );
     Assertions.assertEquals(
-        UserDataset.getDataset().size(),
+        DATASET.size(),
         inPage.getTotalSize() + outPage.getTotalSize()
     );
 
@@ -252,7 +291,7 @@ class JpaWrapperTest {
     allQuery.setFilter("gender IN 'MALE,FEMALE,UNKNOWN'");
     Page<User> allPage = userDAO.query(allQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().size(),
+        DATASET.size(),
         allPage.getTotalSize()
     );
 
@@ -272,7 +311,7 @@ class JpaWrapperTest {
     eqQuery.setFilter(String.format("age EQ %s", exceptNumber));
     Page<User> eqPage = userDAO.query(eqQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> Objects.equals(user.getAge(), 18))
             .count(),
         eqPage.getTotalSize()
@@ -282,13 +321,13 @@ class JpaWrapperTest {
     neQuery.setFilter(String.format("age NE %s", exceptNumber));
     Page<User> nePage = userDAO.query(neQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> !Objects.equals(user.getAge(), 18))
             .count(),
         nePage.getTotalSize()
     );
     Assertions.assertEquals(
-        UserDataset.getDataset().size(),
+        DATASET.size(),
         eqPage.getTotalSize() + nePage.getTotalSize()
     );
 
@@ -296,7 +335,7 @@ class JpaWrapperTest {
     gtQuery.setFilter(String.format("age GT %s", exceptNumber));
     Page<User> gtPage = userDAO.query(gtQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> user.getAge() > 18)
             .count(),
         gtPage.getTotalSize()
@@ -306,7 +345,7 @@ class JpaWrapperTest {
     ltQuery.setFilter(String.format("age LT %s", exceptNumber));
     Page<User> ltPage = userDAO.query(ltQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> user.getAge() < 18)
             .count(),
         ltPage.getTotalSize()
@@ -316,13 +355,13 @@ class JpaWrapperTest {
     geQuery.setFilter(String.format("age GE %s", exceptNumber));
     Page<User> gePage = userDAO.query(geQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> user.getAge() >= 18)
             .count(),
         gePage.getTotalSize()
     );
     Assertions.assertEquals(
-        UserDataset.getDataset().size(),
+        DATASET.size(),
         gePage.getTotalSize() + ltPage.getTotalSize()
     );
 
@@ -330,13 +369,13 @@ class JpaWrapperTest {
     leQuery.setFilter(String.format("age LE %s", exceptNumber));
     Page<User> lePage = userDAO.query(leQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> user.getAge() <= 18)
             .count(),
         lePage.getTotalSize()
     );
     Assertions.assertEquals(
-        UserDataset.getDataset().size(),
+        DATASET.size(),
         lePage.getTotalSize() + gtPage.getTotalSize()
     );
   }
@@ -344,7 +383,7 @@ class JpaWrapperTest {
   @Test
   void query_instant() {
     // yyyy-MM-ddTHH:mm:ss
-    Instant exceptDateTime = UserDataset.getDataset().parallelStream()
+    Instant exceptDateTime = DATASET.parallelStream()
         .findAny()
         .get()
         .getCreateTime();
@@ -356,7 +395,7 @@ class JpaWrapperTest {
     ));
     Page<User> eqPage = userDAO.query(eqQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> Objects.equals(user.getCreateTime(), exceptDateTime))
             .count(),
         eqPage.getTotalSize()
@@ -369,7 +408,7 @@ class JpaWrapperTest {
     ));
     Page<User> nePage = userDAO.query(neQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> !Objects.equals(user.getCreateTime(), exceptDateTime))
             .count(),
         nePage.getTotalSize()
@@ -382,7 +421,7 @@ class JpaWrapperTest {
     ));
     Page<User> gtPage = userDAO.query(gtQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> user.getCreateTime().isAfter(exceptDateTime))
             .count(),
         gtPage.getTotalSize()
@@ -394,7 +433,7 @@ class JpaWrapperTest {
     ));
     Page<User> ltPage = userDAO.query(ltQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> user.getCreateTime().isBefore(exceptDateTime))
             .count(),
         ltPage.getTotalSize()
@@ -406,7 +445,7 @@ class JpaWrapperTest {
     ));
     Page<User> gePage = userDAO.query(geQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> !user.getCreateTime().isBefore(exceptDateTime))
             .count(),
         gePage.getTotalSize()
@@ -418,7 +457,7 @@ class JpaWrapperTest {
     ));
     Page<User> lePage = userDAO.query(leQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> !user.getCreateTime().isAfter(exceptDateTime))
             .count(),
         lePage.getTotalSize()
@@ -428,7 +467,7 @@ class JpaWrapperTest {
   @Test
   void query_datetime() {
     // yyyy-MM-ddTHH:mm:ss
-    LocalDateTime exceptDateTime = UserDataset.getDataset().parallelStream()
+    LocalDateTime exceptDateTime = DATASET.parallelStream()
         .filter(user -> Objects.nonNull(user.getDeleteDateTime()))
         .findAny()
         .get().getDeleteDateTime();
@@ -440,7 +479,7 @@ class JpaWrapperTest {
     ));
     Page<User> eqPage = userDAO.query(eqQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> Objects.nonNull(user.getDeleteDateTime()))
             .filter(user -> Objects.equals(user.getDeleteDateTime(), exceptDateTime))
             .count(),
@@ -454,7 +493,7 @@ class JpaWrapperTest {
     ));
     Page<User> nePage = userDAO.query(neQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> Objects.nonNull(user.getDeleteDateTime()))
             .filter(user -> !Objects.equals(user.getDeleteDateTime(), exceptDateTime))
             .count(),
@@ -468,7 +507,7 @@ class JpaWrapperTest {
     ));
     Page<User> gtPage = userDAO.query(gtQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> Objects.nonNull(user.getDeleteDateTime()))
             .filter(user -> user.getDeleteDateTime().isAfter(exceptDateTime))
             .count(),
@@ -481,7 +520,7 @@ class JpaWrapperTest {
     ));
     Page<User> ltPage = userDAO.query(ltQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> Objects.nonNull(user.getDeleteDateTime()))
             .filter(user -> user.getDeleteDateTime().isBefore(exceptDateTime))
             .count(),
@@ -494,7 +533,7 @@ class JpaWrapperTest {
     ));
     Page<User> gePage = userDAO.query(geQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> Objects.nonNull(user.getDeleteDateTime()))
             .filter(user -> !user.getDeleteDateTime().isBefore(exceptDateTime))
             .count(),
@@ -507,7 +546,7 @@ class JpaWrapperTest {
     ));
     Page<User> lePage = userDAO.query(leQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> Objects.nonNull(user.getDeleteDateTime()))
             .filter(user -> !user.getDeleteDateTime().isAfter(exceptDateTime))
             .count(),
@@ -518,7 +557,7 @@ class JpaWrapperTest {
   @Test
   void query_date() {
     // yyyy-MM-dd
-    LocalDate exceptDate = UserDataset.getDataset().parallelStream()
+    LocalDate exceptDate = DATASET.parallelStream()
         .filter(user -> Objects.nonNull(user.getDeleteDate()))
         .findAny()
         .get().getDeleteDate();
@@ -530,7 +569,7 @@ class JpaWrapperTest {
     ));
     Page<User> eqPage = userDAO.query(eqQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> Objects.nonNull(user.getDeleteDate()))
             .filter(user -> Objects.equals(user.getDeleteDate(), exceptDate))
             .count(),
@@ -544,7 +583,7 @@ class JpaWrapperTest {
     ));
     Page<User> nePage = userDAO.query(neQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> Objects.nonNull(user.getDeleteDate()))
             .filter(user -> !Objects.equals(user.getDeleteDate(), exceptDate))
             .count(),
@@ -558,7 +597,7 @@ class JpaWrapperTest {
     ));
     Page<User> gtPage = userDAO.query(gtQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> Objects.nonNull(user.getDeleteDate()))
             .filter(user -> user.getDeleteDate().isAfter(exceptDate))
             .count(),
@@ -571,7 +610,7 @@ class JpaWrapperTest {
     ));
     Page<User> ltPage = userDAO.query(ltQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> Objects.nonNull(user.getDeleteDate()))
             .filter(user -> user.getDeleteDate().isBefore(exceptDate))
             .count(),
@@ -584,7 +623,7 @@ class JpaWrapperTest {
     ));
     Page<User> gePage = userDAO.query(geQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> Objects.nonNull(user.getDeleteDate()))
             .filter(user -> !user.getDeleteDate().isBefore(exceptDate))
             .count(),
@@ -597,7 +636,7 @@ class JpaWrapperTest {
     ));
     Page<User> lePage = userDAO.query(leQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> Objects.nonNull(user.getDeleteDate()))
             .filter(user -> !user.getDeleteDate().isAfter(exceptDate))
             .count(),
@@ -608,7 +647,7 @@ class JpaWrapperTest {
   @Test
   void query_time() {
     // HH:mm:ss
-    LocalTime exceptTime = UserDataset.getDataset().parallelStream()
+    LocalTime exceptTime = DATASET.parallelStream()
         .filter(user -> Objects.nonNull(user.getDeleteTime()))
         .findAny()
         .get().getDeleteTime();
@@ -620,7 +659,7 @@ class JpaWrapperTest {
     ));
     Page<User> eqPage = userDAO.query(eqQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> Objects.nonNull(user.getDeleteTime()))
             .filter(user -> Objects.equals(user.getDeleteTime(), exceptTime))
             .count(),
@@ -634,7 +673,7 @@ class JpaWrapperTest {
     ));
     Page<User> nePage = userDAO.query(neQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> Objects.nonNull(user.getDeleteTime()))
             .filter(user -> !Objects.equals(user.getDeleteTime(), exceptTime))
             .count(),
@@ -648,7 +687,7 @@ class JpaWrapperTest {
     ));
     Page<User> gtPage = userDAO.query(gtQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> Objects.nonNull(user.getDeleteTime()))
             .filter(user -> user.getDeleteTime().isAfter(exceptTime))
             .count(),
@@ -661,7 +700,7 @@ class JpaWrapperTest {
     ));
     Page<User> ltPage = userDAO.query(ltQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> Objects.nonNull(user.getDeleteTime()))
             .filter(user -> user.getDeleteTime().isBefore(exceptTime))
             .count(),
@@ -674,7 +713,7 @@ class JpaWrapperTest {
     ));
     Page<User> gePage = userDAO.query(geQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> Objects.nonNull(user.getDeleteTime()))
             .filter(user -> !user.getDeleteTime().isBefore(exceptTime))
             .count(),
@@ -687,7 +726,7 @@ class JpaWrapperTest {
     ));
     Page<User> lePage = userDAO.query(leQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> Objects.nonNull(user.getDeleteTime()))
             .filter(user -> !user.getDeleteTime().isAfter(exceptTime))
             .count(),
@@ -701,7 +740,7 @@ class JpaWrapperTest {
     eqQuery.setFilter("deleted EQ true");
     Page<User> eqPage = userDAO.query(eqQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> Objects.equals(user.getDeleted(), true))
             .count(),
         eqPage.getTotalSize()
@@ -711,7 +750,7 @@ class JpaWrapperTest {
     neQuery.setFilter("deleted NE false");
     Page<User> nePage = userDAO.query(neQuery);
     Assertions.assertEquals(
-        UserDataset.getDataset().parallelStream()
+        DATASET.parallelStream()
             .filter(user -> !Objects.equals(user.getDeleted(), false))
             .count(),
         nePage.getTotalSize()
@@ -723,135 +762,3 @@ class JpaWrapperTest {
   }
 }
 
-@Getter
-@Setter
-@Accessors(chain = true)
-@Entity
-@Table(name = "t_user")
-class User {
-
-  @Id
-  @Column
-  private String id;
-  @Column(nullable = false)
-  private String name;
-  @Column
-  private Integer age;
-  @Column(precision = 13, scale = 9)
-  private BigDecimal credit;
-  @Column(updatable = false)
-  private GenderEnum gender;
-  @Column(updatable = false)
-  private Instant createTime;
-  @Column
-  private LocalDateTime deleteDateTime;
-  @Column
-  private Boolean deleted;
-  @Column
-  private LocalDate deleteDate;
-  @Column
-  private LocalTime deleteTime;
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    User user = (User) o;
-    return Objects.equals(id, user.id);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(id);
-  }
-}
-
-enum GenderEnum {
-  MALE,
-  FEMALE,
-  UNKNOWN,
-  ;
-}
-
-@Singleton
-class UserDataset {
-
-  private static final List<User> DATASET = new ArrayList<>();
-
-  static {
-    for (int i = 0; i < 1000; i++) {
-      User user = new User();
-      user.setId(IdUtil.getSnowflakeNextIdStr());
-      user.setName(RandomUtil.randomString(6));
-      user.setAge(RandomUtil.randomInt(6, 60));
-      user.setCredit(RandomUtil.randomBigDecimal(new BigDecimal("100.000")));
-      user.setGender(RandomUtil.randomEle(GenderEnum.values()));
-      user.setCreateTime(Instant.EPOCH.plus(
-          RandomUtil.randomLong(365 * 30, 365 * 300),
-          ChronoUnit.DAYS
-      ));
-      user.setDeleteDateTime(RandomUtil.randomBoolean() ?
-          LocalDateTime.of(
-              RandomUtil.randomInt(2000, 2999),
-              RandomUtil.randomInt(1, 11),
-              RandomUtil.randomInt(1, 27),
-              RandomUtil.randomInt(1, 11),
-              RandomUtil.randomInt(1, 59),
-              RandomUtil.randomInt(1, 59)
-          ) : null
-      );
-      user.setDeleted(user.getDeleteDateTime() != null);
-      user.setDeleteDate(
-          user.getDeleteDateTime() == null
-              ? null
-              : user.getDeleteDateTime().toLocalDate()
-      );
-      user.setDeleteTime(
-          user.getDeleteDateTime() == null
-              ? null
-              : user.getDeleteDateTime().toLocalTime()
-      );
-      DATASET.add(user);
-    }
-  }
-
-  public static List<User> getDataset() {
-    return new ArrayList<>(DATASET);
-  }
-
-  @Inject
-  protected transient UserRepository repository;
-
-  @EventListener
-  public void init(ApplicationStartupEvent event) {
-    repository.saveAll(DATASET);
-  }
-}
-
-@Repository
-interface UserRepository extends JpaRepository<User, String> {
-
-}
-
-@Repository
-class UserDAO extends BaseJpaWrapper<User, QUser, UserRepository> {
-
-
-  protected UserDAO(UserRepository dao) {
-    super(dao);
-  }
-
-  @Override
-  protected QUser getQuery() {
-    return QUser.user;
-  }
-
-  @Override
-  protected Class<User> getBeanClass() {
-    return User.class;
-  }
-}
