+++ micronaut-tao/src/test/java/io/github/xezzon/tao/jpa/user/GenderEnum.java
@@ -1,0 +1,11 @@
+package io.github.xezzon.tao.jpa.user;
+
+/**
+ * @author xezzon
+ */
+public enum GenderEnum {
+  MALE,
+  FEMALE,
+  UNKNOWN,
+  ;
+}
