+++ micronaut-tao/src/test/java/io/github/xezzon/tao/jpa/user/UserDAO.java
@@ -1,0 +1,26 @@
+package io.github.xezzon.tao.jpa.user;
+
+import io.github.xezzon.tao.jpa.BaseJpaWrapper;
+import io.micronaut.data.annotation.Repository;
+
+/**
+ * @author xezzon
+ */
+@Repository
+public class UserDAO extends BaseJpaWrapper<User, QUser, UserRepository> {
+
+
+  protected UserDAO(UserRepository dao) {
+    super(dao);
+  }
+
+  @Override
+  protected QUser getQuery() {
+    return QUser.user;
+  }
+
+  @Override
+  protected Class<User> getBeanClass() {
+    return User.class;
+  }
+}
