+++ micronaut-tao/src/test/java/io/github/xezzon/tao/jpa/user/UserRepository.java
@@ -1,0 +1,12 @@
+package io.github.xezzon.tao.jpa.user;
+
+import io.micronaut.data.annotation.Repository;
+import io.micronaut.data.jpa.repository.JpaRepository;
+
+/**
+ * @author xezzon
+ */
+@Repository
+public interface UserRepository extends JpaRepository<User, String> {
+
+}
